# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vertretungsplan API
  description: ''
  contact:
    email: support@schulzentrum.de
  license:
    name: Open Source License 3.0
    url: https://opensource.org/license/osl-3-0
  version: 0.0.5
servers:
  - url: http://localhost:80
  - url: https://secure-gateway.schulzentrum.de/api/v1/timetable
...
paths:
...
  /metadata/current-period/{timestamp}:
    parameters:
      - $ref: '#/components/parameters/TimestampPath'
    get:
      tags:
        - metadata
      summary: Returns the current active lesson period based on provided timestamp
      responses:
        '200':
          description: Current active lesson period
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Period'
...
  /plans/daily/{date}/classes:
    parameters:
      - $ref: '#/components/parameters/DatePath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and all classes
      responses:
        '200':
          description: Timetable for all classes on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/daily/{date}/classes/{id}:
    parameters:
      - $ref: '#/components/parameters/DatePath'
      - $ref: '#/components/parameters/ClassIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and class
      responses:
        '200':
          description: Timetable for the specified class on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date or class ID format
        '404':
          description: Class not found
        '500':
          description: Internal Server Error
...
  /plans/daily/today/classes:
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and all classes
      responses:
        '200':
          description: Timetable for all classes for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '500':
          description: Internal Server Error
  /plans/daily/today/classes/{id}:
    parameters:
      - $ref: '#/components/parameters/ClassIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and a specific class
      responses:
        '200':
          description: Timetable for the specified class for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          description: Class not found
        '500':
          description: Internal Server Error
...
  /plans/daily/{date}/teachers:
    parameters:
      - $ref: '#/components/parameters/DatePath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and all teachers
      responses:
        '200':
          description: Timetable for all teachers on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/daily/{date}/teachers/{id}:
    parameters:
      - $ref: '#/components/parameters/DatePath'
      - $ref: '#/components/parameters/TeacherIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and teacher
      responses:
        '200':
          description: Timetable for the specified teacher on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date or teacher ID format
        '404':
          description: Teacher not found
        '500':
          description: Internal Server Error
...
  /plans/daily/today/teachers:
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and all teachers
      responses:
        '200':
          description: Timetable for all teachers for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '500':
          description: Internal Server Error
  /plans/daily/today/teachers/{id}:
    parameters:
      - $ref: '#/components/parameters/TeacherIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and a specific teacher
      responses:
        '200':
          description: Timetable for the specified teacher for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          description: Teacher not found
        '500':
          description: Internal Server Error
...
  /plans/daily/{date}/rooms:
    parameters:
      - $ref: '#/components/parameters/DatePath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and all rooms
      responses:
        '200':
          description: Timetable for all rooms on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/daily/{date}/rooms/{roomID}:
    parameters:
      - $ref: '#/components/parameters/DatePath'
      - $ref: '#/components/parameters/RoomIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for a specific date and room
      responses:
        '200':
          description: Timetable for the specified room on the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          description: Invalid date or room ID format
        '404':
          description: Room not found
        '500':
          description: Internal Server Error
...
  /plans/daily/today/rooms:
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and all rooms
      responses:
        '200':
          description: Timetable for all rooms for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '500':
          description: Internal Server Error
  /plans/daily/today/rooms/{roomID}:
    parameters:
      - $ref: '#/components/parameters/RoomIdPath'
    get:
      tags:
        - plan-daily
      summary: Returns timetable for today and a specific room
      responses:
        '200':
          description: Timetable for the specified room for today
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          description: Room not found
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/classes:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and all classes
      responses:
        '200':
          description: Weekly timetable for all classes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/classes/{id}:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
      - $ref: '#/components/parameters/ClassIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and class
      responses:
        '200':
          description: Weekly timetable for the specified class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date or class ID format
        '404':
          description: Class not found
        '500':
          description: Internal Server Error
...
  /plans/weekly/today/classes/{id}:
    parameters:
      - $ref: '#/components/parameters/ClassIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for this week and class
      responses:
        '200':
          description: Weekly timetable for the specified class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid class ID format
        '404':
          description: Class not found
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/teachers:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and all teachers
      responses:
        '200':
          description: Weekly timetable for all teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/teachers/{id}:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
      - $ref: '#/components/parameters/TeacherIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and teacher
      responses:
        '200':
          description: Weekly timetable for the specified teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date or teacher ID format
        '404':
          description: Teacher not found
        '500':
          description: Internal Server Error
  /plans/weekly/today/teachers:
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for this week and all teachers
      responses:
        '200':
          description: Weekly timetable for all teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '500':
          description: Internal Server Error
  /plans/weekly/today/teachers/{id}:
    parameters:
      - $ref: '#/components/parameters/TeacherIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for this week and teacher
      responses:
        '200':
          description: Weekly timetable for the specified teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid teacher ID format
        '404':
          description: Teacher not found
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/rooms:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and all rooms
      responses:
        '200':
          description: Weekly timetable for all rooms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date format
        '500':
          description: Internal Server Error
  /plans/weekly/{date}/rooms/{roomID}:
    parameters:
      - $ref: '#/components/parameters/WeekDatePath'
      - $ref: '#/components/parameters/RoomIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for a week and room
      responses:
        '200':
          description: Weekly timetable for the specified room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid date or room ID format
        '404':
          description: Room not found
        '500':
          description: Internal Server Error
  /plans/weekly/today/rooms:
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for this week and all rooms
      responses:
        '200':
          description: Weekly timetable for all rooms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '500':
          description: Internal Server Error
  /plans/weekly/today/rooms/{roomID}:
    parameters:
      - $ref: '#/components/parameters/RoomIdPath'
    get:
      tags:
        - plan-weekly
      summary: Returns timetable for this week and room
      responses:
        '200':
          description: Weekly timetable for the specified room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyPlanResponse'
        '400':
          description: Invalid room ID format
        '404':
          description: Room not found
        '500':
          description: Internal Server Error
...
components:
  parameters:
    DatePath:
      name: date
      description: Date for requested timetable
      in: path
      required: true
      schema:
        type: string
        format: date
    WeekDatePath:
      name: date
      description: Date that is inside requested week timetable
      in: path
      required: true
      schema:
        type: string
        format: date
    ClassIdPath:
      name: id
      description: ID of a specific class
      in: path
      required: true
      schema:
        type: integer
    TeacherIdPath:
      name: id
      description: ID of a specific teacher
      in: path
      required: true
      schema:
        type: integer
    RoomIdPath:
      name: roomID
      description: ID of a specific room
      in: path
      required: true
      schema:
        type: integer
    TimestampPath:
      name: timestamp
      description: Timestamp to check for active lesson period without timezone info
      in: path
      required: true
      schema:
        type: string
        format: date-time
      example: '2023-10-05T10:15:00'
  schemas:
    Teacher:
      type: object
      required:
        - id
        - shortName
      properties:
        id:
          type: integer
          examples:
            - 101
        shortName:
          type: string
          description: Unique initials or short code of the teacher
          examples:
            - SMI
    Room:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          examples:
            - 50
        name:
          type: string
          description: Room number or name
          examples:
            - '3.12'
        building:
          type: string
          examples:
            - Main Building
    SchoolClass:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          examples:
            - 12
        name:
          type: string
          description: Grade and identifier
          examples:
            - 10A
    Subject:
      type: object
      required:
        - id
        - shortName
      properties:
        id:
          type: integer
          examples:
            - 5
        shortName:
          type: string
          examples:
            - MA
        longName:
          type: string
          examples:
            - Mathematics
    Period:
      type: object
      required:
        - number
        - startTime
        - endTime
      properties:
        number:
          type: integer
          description: The lesson number
          examples:
            - 1
        startTime:
          type: string
          pattern: ^([01]\d|2[0-3]):?([0-5]\d)$
          examples:
            - '08:00'
        endTime:
          type: string
          pattern: ^([01]\d|2[0-3]):?([0-5]\d)$
          examples:
            - '08:45'
...
    TimetableEntry:
      type: object
      required:
        - id
        - date
        - period
        - status
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        period:
          $ref: '#/components/schemas/Period'
        subject:
          $ref: '#/components/schemas/Subject'
        teacher:
          $ref: '#/components/schemas/Teacher'
        room:
          $ref: '#/components/schemas/Room'
        classes:
          type: array
          items:
            $ref: '#/components/schemas/SchoolClass'
        status:
          type: string
          description: Current state of the lesson
          enum:
            - scheduled
            - substituted
            - cancelled
            - time_change
            - room_change
        substitutionInfo:
          type: object
          description: Details if the lesson is not 'scheduled'
          properties:
            originalTeacher:
              $ref: '#/components/schemas/Teacher'
            originalRoom:
              $ref: '#/components/schemas/Room'
            changeReason:
              type: string
              examples:
                - Illness
            notes:
              type: string
              examples:
                - Work on textbook page 45
...
    PlanResponse:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TimetableEntry'
    WeeklyPlanResponse:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
          description: The timestamp when the response was generated.
        entries:
          type: object
          additionalProperties:
            type: object
            description: Timetable entries grouped by date.
            additionalProperties:
              type: array
              items:
                $ref: '#/components/schemas/TimetableEntry'